requirements.txt
google-genai
python-dotenv
langchain
streamlit
pypdf
sqlalchemy
psycopg2
pandas
supabase
pinecone